feat: decouple Konva from core (#54)
Konva has been decoupled from the core entirely. It is now possible to supply the project with custom implementations of the `Scene` interface.
Because of that, scene files need to export a special `SceneDescription` object instead of a simple generator function.

The preferred way of doing it is by using dedicated "make" functions that generate description objects for specific scenes.
For example this scene:
```ts
export default function* example(scene: Scene): ThreadGenerator {
  yield* scene.transition();
  // perform animation
}
```
should now be written as:
```ts
export default makeKonvaScene(function* example(view) {
  yield* view.transition();
  // perform animation
});
```

The core provides an abstract implementation of the `Scene` interface called `GeneratorScene`.
It takes care of all the logic related to generators, life cycles, caching, etc. And can be extended to easily provide custom rendering logic. See `KonvaScene` for an example of extending the `GeneratorScene`.

Alternatively, the `Scene` interface can be implemented from the ground up allowing for even more flexibility.

For the time being, features related to Konva are kept in the core repo.
After #49 they will be extracted to a separate package.

BREAKING CHANGE: change the type exported by scene files

Scene files need to export a special `SceneDescription` object instead of a simple generator function.

Closes: #31