feat: added useContext and useContextAfter hooks (#63)
The `useContext` hooks give direct access the the context to allow for
transforming, clipping, and drawing to the canvas directly. The
`useContext` hook runs before the scene is rendered, and so is most
appropriate for altering the render through transforms and clipping. The
`useContextAfter` hook runs after the scene is rendered, and so is most
appropriate for altering the image data or drawing to the canvas.

Fixes #58