feat: added file type and quality options to rendering panel (#50)
The quality setting won't effect PNG, as I believe PNG files have to use
for a quality setting. The setting input will then be inert when PNG,
the default, is used. There are other design options available, but this
is the simplest option that avoids frustration for the user. The
image/webp option will not work on Safari, and there is no check for
this occurance. The best option would be a proactive error message.

Closes #24