feat: add default renderer
This is the initial draft of the new renderer described in #56.
Yoga has been dropped in favor of the browser's built-in layout engine.
The DOM API is mostly the same as Yoga's but requires significantly less effort to set up.
It also opens the door for utilizing CSS in the future.

This PR also contains an initial implementation of properties from #58.
The name has been changed from atoms to signals to match the convention of other libraries.