fix: the resolution fields in Rendering no longer reset each other
The `getSize` method did not cause the React component to observe the
result, as you would get with `useState`. This caused the component to
never receive the new values when after changing them, and upon setting
the other value (whichever comes second), it would re-set the first
value to its original setting. This is the quickest possible fix.