feat(examples): set base url for generated examples
Setting Vite's `base` option to the path at which examples
are statically hosted in the documentation allows for all file
types to be loaded.
Without this it is not possible to use Components like `Image`
or `Video` in any example that should be featured in the
documentation.
Keep in mind the `base` option has been provided in
the build script and not the `vite.config.ts` because
we don't want to mess with the serving for the example
editor.

**Explanation**
Lets imagine we want to import an image from as follows:

```ts
import someSvg from '@motion-canvas/examples/assets/some.svg';
```

which we then provide to an `Image.src` like so:

```tsx
<Image src={someSvg} />
```

The example would work fine in the examples editor when we serve them
but would not work in the documentation. This is due to the fact that
when serving the examples everything in the examples root directory will be
statically served at `/` including the `assets` directory (`/assets`).

However this is not the case for the build examples in the documentation
since they are located in a sub-path (`/examples`).
When bundling Vite doesn't know this and assumes that everything will be
served under `/` and resolves the import path to `/some.svg`.
Setting the `base` option lets Vite know that the result will be located
under `/examples` and the bundled resouces resolve to `/examples/some.svg`
where they then can be located correctly by the browser.