fix(vite-plugin): ignore query param in devserver (#351)
The previous implementation tested the full url, including query
parameters against the '/' path. Any query parameter in the URL would
cause a 404 instead of loading the editor.

Parsing the request URL with the URL class gives us access to the
pathname, which does not include query parameters and can be safely
compared.