feat: new playback architecture
This PR decouples the `Project` and `Player` classes to provide a more
sustainable architecture moving forward.

Rendering has been extracted from the `Player` and is now a standalone process
handled by the `Renderer` class. The `PlaybackManager` serves as the base for
both the `Player` and `Renderer`. In the future, it will be used to handle the
presentation process as well.

The `Renderer` can outsource the actual exporting to any class implementing the
`Exporter` interface. This is the first step towards custom exporters.

Meta files now use a runtime representation that takes care of serialization,
parsing, and notifying others about changes. Each field in the file is
represented by a separate class corresponding to the type of the value stored.
In the future, the editor will generate fields automatically, based on this
representation. This will let features such as Editable Signals and Custom
Exporters display custom settings in the editor, without modifying its source
code.

BREAKING CHANGE: `makeProject` no longer accepts some settings.

Settings such as `background` and `audioOffset` are now stored in the project
meta file.

Closes: #166