build: remove `baseUrl` option (#791)
`baseUrl` is used to declare the base directory for resolving non-relative
module paths. Our tooling is not configured to support that feature and so
the option is not necessary.

Some IDE consider the presence of this option as a hint to use absolute
paths for auto imports. Removing it should fix that issue.