docs: change license of ffmpeg exporter to GPLv3
The ffmpeg exporter package uses ffmpeg, unsurprisingly, and parts of ffmpeg are licensed under the GPL license.
This requires us to license the ffmpeg exporter with the GPL license as well.
As this exporter is independent from the rest of Motion Canvas, the rest of the packages can keep the MIT license.